refactor: YataiService Store Abstraction (#1541)
* fix: makefile sphinx deps

* docs: added batch_serving docs and linked to it in quickstart

* refactor database operations into db class

* remove old db code

* migrate store related functionality to yatai/db/stores module

* makefile + sources typo

* refactor to static classes

* finish static refactor

* fix yatai_service_impl

* fix bad patch in test_usage_stats

* patch metadatastore property naming bug

* lint + format

* more lint + fmt

* lint + format (again)
