+++ bentoml/yatai/validator/deployment_pb_validator.py
@@ -16,7 +16,7 @@ from cerberus import Validator
 
 from bentoml.exceptions import InvalidArgument
 from bentoml.utils import ProtoMessageToDict
-from bentoml.yatai.label_store import _validate_labels
+from bentoml.yatai.db.stores.label import _validate_labels
 from bentoml.yatai.deployment.azure_functions.constants import (
     AZURE_FUNCTIONS_PREMIUM_PLAN_SKUS,
     AZURE_FUNCTIONS_AUTH_LEVELS,
